[exporterhelper] Move NewMetricsRequest to exporterhelper/internal
