+++ exporter/exporterhelper/internal/new_request.go
@@ -11,6 +11,7 @@ import (
 	"go.opentelemetry.io/collector/exporter"
 	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/request"
 	"go.opentelemetry.io/collector/pdata/plog"
+	"go.opentelemetry.io/collector/pdata/pmetric"
 	"go.opentelemetry.io/collector/pdata/ptrace"
 	"go.opentelemetry.io/collector/pipeline"
 	"go.uber.org/zap"
@@ -115,6 +116,59 @@ func newConsumeTraces(converter request.RequestConverterFunc[ptrace.Traces], be 
 		if err != nil {
 			logger.Error("Failed to convert traces. Dropping data.",
 				zap.Int("dropped_spans", td.SpanCount()),
+				zap.Error(err))
+			return consumererror.NewPermanent(err)
+		}
+		return be.Send(ctx, req)
+	}
+}
+
+type metricsExporter struct {
+	*BaseExporter
+	consumer.Metrics
+}
+
+// NewMetricsRequest creates a new metrics exporter based on a custom MetricsConverter and Sender.
+// Experimental: This API is at the early stage of development and may change without backward compatibility
+// until https://github.com/open-telemetry/opentelemetry-collector/issues/8122 is resolved.
+func NewMetricsRequest(
+	_ context.Context,
+	set exporter.Settings,
+	converter request.RequestConverterFunc[pmetric.Metrics],
+	pusher request.RequestConsumeFunc,
+	options ...Option,
+) (exporter.Metrics, error) {
+	if set.Logger == nil {
+		return nil, errNilLogger
+	}
+
+	if converter == nil {
+		return nil, errNilMetricsConverter
+	}
+
+	if pusher == nil {
+		return nil, errNilConsumeRequest
+	}
+
+	be, err := NewBaseExporter(set, pipeline.SignalMetrics, pusher, options...)
+	if err != nil {
+		return nil, err
+	}
+
+	mc, err := consumer.NewMetrics(newConsumeMetrics(converter, be, set.Logger), be.ConsumerOptions...)
+	if err != nil {
+		return nil, err
+	}
+
+	return &metricsExporter{BaseExporter: be, Metrics: mc}, nil
+}
+
+func newConsumeMetrics(converter request.RequestConverterFunc[pmetric.Metrics], be *BaseExporter, logger *zap.Logger) consumer.ConsumeMetricsFunc {
+	return func(ctx context.Context, md pmetric.Metrics) error {
+		req, err := converter(ctx, md)
+		if err != nil {
+			logger.Error("Failed to convert metrics. Dropping data.",
+				zap.Int("dropped_data_points", md.DataPointCount()),
 				zap.Error(err))
 			return consumererror.NewPermanent(err)
 		}
