+++ exporter/exporterhelper/metrics_test.go
@@ -71,46 +71,15 @@ func TestMetrics_NilLogger(t *testing.T) {
 	require.Equal(t, errNilLogger, err)
 }
 
-func TestMetricsRequest_NilLogger(t *testing.T) {
-	me, err := NewMetricsRequest(context.Background(), exporter.Settings{}, requesttest.RequestFromMetricsFunc(nil), sendertest.NewNopSenderFunc[Request]())
-	require.Nil(t, me)
-	require.Equal(t, errNilLogger, err)
-}
-
 func TestMetrics_NilPushMetricsData(t *testing.T) {
 	me, err := NewMetrics(context.Background(), exportertest.NewNopSettings(exportertest.NopType), &fakeMetricsConfig, nil)
 	require.Nil(t, me)
 	require.Equal(t, errNilPushMetrics, err)
 }
 
-func TestMetricsRequest_NilMetricsConverter(t *testing.T) {
-	me, err := NewMetricsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType), nil, sendertest.NewNopSenderFunc[Request]())
-	require.Nil(t, me)
-	require.Equal(t, errNilMetricsConverter, err)
-}
-
-func TestMetricsRequest_NilPushMetricsData(t *testing.T) {
-	me, err := NewMetricsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType), requesttest.RequestFromMetricsFunc(nil), nil)
-	require.Nil(t, me)
-	require.Equal(t, errNilConsumeRequest, err)
-}
-
 func TestMetrics_Default(t *testing.T) {
 	md := pmetric.NewMetrics()
 	me, err := NewMetrics(context.Background(), exportertest.NewNopSettings(exportertest.NopType), &fakeMetricsConfig, newPushMetricsData(nil))
-	require.NoError(t, err)
-	assert.NotNil(t, me)
-
-	assert.Equal(t, consumer.Capabilities{MutatesData: false}, me.Capabilities())
-	assert.NoError(t, me.Start(context.Background(), componenttest.NewNopHost()))
-	assert.NoError(t, me.ConsumeMetrics(context.Background(), md))
-	assert.NoError(t, me.Shutdown(context.Background()))
-}
-
-func TestMetricsRequest_Default(t *testing.T) {
-	md := pmetric.NewMetrics()
-	me, err := NewMetricsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType),
-		requesttest.RequestFromMetricsFunc(nil), sendertest.NewNopSenderFunc[Request]())
 	require.NoError(t, err)
 	assert.NotNil(t, me)
 
@@ -129,16 +98,6 @@ func TestMetrics_WithCapabilities(t *testing.T) {
 	assert.Equal(t, capabilities, me.Capabilities())
 }
 
-func TestMetricsRequest_WithCapabilities(t *testing.T) {
-	capabilities := consumer.Capabilities{MutatesData: true}
-	me, err := NewMetricsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType),
-		requesttest.RequestFromMetricsFunc(nil), sendertest.NewNopSenderFunc[Request](), WithCapabilities(capabilities))
-	require.NoError(t, err)
-	assert.NotNil(t, me)
-
-	assert.Equal(t, capabilities, me.Capabilities())
-}
-
 func TestMetrics_Default_ReturnError(t *testing.T) {
 	md := pmetric.NewMetrics()
 	want := errors.New("my_error")
@@ -148,25 +107,6 @@ func TestMetrics_Default_ReturnError(t *testing.T) {
 	require.Equal(t, want, me.ConsumeMetrics(context.Background(), md))
 }
 
-func TestMetricsRequest_Default_ConvertError(t *testing.T) {
-	md := pmetric.NewMetrics()
-	want := errors.New("convert_error")
-	me, err := NewMetricsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType),
-		requesttest.RequestFromMetricsFunc(want), sendertest.NewNopSenderFunc[Request]())
-	require.NoError(t, err)
-	require.NotNil(t, me)
-	require.Equal(t, consumererror.NewPermanent(want), me.ConsumeMetrics(context.Background(), md))
-}
-
-func TestMetricsRequest_Default_ExportError(t *testing.T) {
-	md := pmetric.NewMetrics()
-	want := errors.New("export_error")
-	me, err := NewMetricsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType),
-		requesttest.RequestFromMetricsFunc(nil), sendertest.NewErrSenderFunc[Request](want))
-	require.NoError(t, err)
-	require.NotNil(t, me)
-	require.Equal(t, want, me.ConsumeMetrics(context.Background(), md))
-}
 
 func TestMetrics_WithPersistentQueue(t *testing.T) {
 	fgOrigReadState := queue.PersistRequestContextOnRead
@@ -279,7 +219,7 @@ func TestMetricsRequest_WithRecordMetrics(t *testing.T) {
 	tt := componenttest.NewTelemetry()
 	t.Cleanup(func() { require.NoError(t, tt.Shutdown(context.Background())) })
 
-	me, err := NewMetricsRequest(context.Background(),
+	me, err := internal.NewMetricsRequest(context.Background(),
 		exporter.Settings{ID: fakeMetricsName, TelemetrySettings: tt.NewTelemetrySettings(), BuildInfo: component.NewDefaultBuildInfo()},
 		requesttest.RequestFromMetricsFunc(nil), sendertest.NewNopSenderFunc[Request]())
 	require.NoError(t, err)
@@ -305,7 +245,7 @@ func TestMetricsRequest_WithRecordMetrics_ExportError(t *testing.T) {
 	tt := componenttest.NewTelemetry()
 	t.Cleanup(func() { require.NoError(t, tt.Shutdown(context.Background())) })
 
-	me, err := NewMetricsRequest(context.Background(),
+	me, err := internal.NewMetricsRequest(context.Background(),
 		exporter.Settings{ID: fakeMetricsName, TelemetrySettings: tt.NewTelemetrySettings(), BuildInfo: component.NewDefaultBuildInfo()},
 		requesttest.RequestFromMetricsFunc(nil), sendertest.NewErrSenderFunc[Request](want))
 	require.NoError(t, err)
@@ -334,7 +274,7 @@ func TestMetricsRequest_WithSpan(t *testing.T) {
 	otel.SetTracerProvider(set.TracerProvider)
 	defer otel.SetTracerProvider(nooptrace.NewTracerProvider())
 
-	me, err := NewMetricsRequest(context.Background(), set, requesttest.RequestFromMetricsFunc(nil), sendertest.NewNopSenderFunc[Request]())
+	me, err := internal.NewMetricsRequest(context.Background(), set, requesttest.RequestFromMetricsFunc(nil), sendertest.NewNopSenderFunc[Request]())
 	require.NoError(t, err)
 	require.NotNil(t, me)
 	checkWrapSpanForMetrics(t, sr, set.TracerProvider.Tracer("test"), me, nil)
@@ -362,7 +302,7 @@ func TestMetricsRequest_WithSpan_ExportError(t *testing.T) {
 	defer otel.SetTracerProvider(nooptrace.NewTracerProvider())
 
 	want := errors.New("my_error")
-	me, err := NewMetricsRequest(context.Background(), set, requesttest.RequestFromMetricsFunc(nil), sendertest.NewErrSenderFunc[Request](want))
+	me, err := internal.NewMetricsRequest(context.Background(), set, requesttest.RequestFromMetricsFunc(nil), sendertest.NewErrSenderFunc[Request](want))
 	require.NoError(t, err)
 	require.NotNil(t, me)
 	checkWrapSpanForMetrics(t, sr, set.TracerProvider.Tracer("test"), me, want)
@@ -381,38 +321,11 @@ func TestMetrics_WithShutdown(t *testing.T) {
 	assert.True(t, shutdownCalled)
 }
 
-func TestMetricsRequest_WithShutdown(t *testing.T) {
-	shutdownCalled := false
-	shutdown := func(context.Context) error { shutdownCalled = true; return nil }
-
-	me, err := NewMetricsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType),
-		requesttest.RequestFromMetricsFunc(nil), sendertest.NewNopSenderFunc[Request](), WithShutdown(shutdown))
-	assert.NotNil(t, me)
-	assert.NoError(t, err)
-
-	assert.NoError(t, me.Start(context.Background(), componenttest.NewNopHost()))
-	assert.NoError(t, me.Shutdown(context.Background()))
-	assert.True(t, shutdownCalled)
-}
-
 func TestMetrics_WithShutdown_ReturnError(t *testing.T) {
 	want := errors.New("my_error")
 	shutdownErr := func(context.Context) error { return want }
 
 	me, err := NewMetrics(context.Background(), exportertest.NewNopSettings(exportertest.NopType), &fakeMetricsConfig, newPushMetricsData(nil), WithShutdown(shutdownErr))
-	assert.NotNil(t, me)
-	assert.NoError(t, err)
-
-	assert.NoError(t, me.Start(context.Background(), componenttest.NewNopHost()))
-	assert.Equal(t, want, me.Shutdown(context.Background()))
-}
-
-func TestMetricsRequest_WithShutdown_ReturnError(t *testing.T) {
-	want := errors.New("my_error")
-	shutdownErr := func(context.Context) error { return want }
-
-	me, err := NewMetricsRequest(context.Background(), exportertest.NewNopSettings(exportertest.NopType),
-		requesttest.RequestFromMetricsFunc(nil), sendertest.NewNopSenderFunc[Request](), WithShutdown(shutdownErr))
 	assert.NotNil(t, me)
 	assert.NoError(t, err)
